autogen_sample/requirements.txt
pyautogen[retrievechat]
docker
chromadb
ipython
flaml[automl]